+++ test/libsolidity/ASTJSONTest.h
@@ -38,7 +38,9 @@ class ASTJSONTest: public TestCase
 {
 public:
 	static std::unique_ptr<TestCase> create(Config const& _config)
-	{ return std::make_unique<ASTJSONTest>(_config.filename); }
+	{
+		return std::make_unique<ASTJSONTest>(_config.filename);
+	}
 	ASTJSONTest(std::string const& _filename);
 
 	TestResult run(std::ostream& _stream, std::string const& _linePrefix = "", bool const _formatted = false) override;
@@ -51,7 +53,6 @@ private:
 		std::string& _result,
 		std::map<std::string, unsigned> const& _sourceIndices,
 		CompilerStack& _compiler,
-		bool _parseOnly,
 		std::string const& _variation,
 		std::ostream& _stream,
 		std::string const& _linePrefix = "",
@@ -64,13 +65,10 @@ private:
 	) const;
 
 	std::vector<std::pair<std::string, std::string>> m_sources;
-	std::string m_expectationLegacy;
 	std::string m_expectationParseOnly;
 	std::string m_astFilename;
 	std::string m_astParseOnlyFilename;
-	std::string m_legacyAstFilename;
 	std::string m_result;
-	std::string m_resultLegacy;
 	std::string m_resultParseOnly;
 };